- spark service implemented

+++ src/main/java/ufrn/imd/engsoft/dao/TweetsDAO.java
@@ -2,28 +2,35 @@ package ufrn.imd.engsoft.dao;
 
 import com.mongodb.DB;
 import com.mongodb.MongoClient;
-import ufrn.imd.engsoft.model.TweetInfo;
 import org.jongo.Jongo;
+import org.jongo.MongoCursor;
+import ufrn.imd.engsoft.model.TweetInfo;
 
 import java.util.List;
 
 /**
  * Created by Felipe on 10/16/15.
  */
-public class TweetsDAO {
-
+public class TweetsDAO
+{
     private static TweetsDAO _instance;
+    private static boolean _dropCollection;
     private String _collectionName;
     private DB _database;
     private Jongo _jongo;
     private MongoClient _mongoClient;
 
-    private TweetsDAO(String _collectionName) {
+    private TweetsDAO(String collectionName)
+    {
         _mongoClient = new MongoClient();
-        _database = _mongoClient.getDB("tweets-db");
+        _database = _mongoClient.getDB("tweets_db");
         _jongo = new Jongo(_database);
-        this._collectionName = _collectionName;
-        dropCollection();
+        _collectionName = collectionName;
+
+        if(_dropCollection)
+        {
+            dropCollection();
+        }
     }
 
     private static synchronized void createInstance (String collectionName)
@@ -34,26 +41,41 @@ public class TweetsDAO {
         }
     }
 
-    public static TweetsDAO getInstance(String collectionName){
-
+    public static TweetsDAO getInstance(String collectionName, boolean dropCollection)
+    {
         if(_instance == null)
         {
+            _dropCollection = dropCollection;
             createInstance (collectionName);
 
         }
         return _instance;
-
     }
 
-    public void save(List<TweetInfo> tweetInfoList){
-        _jongo.getCollection(_collectionName).save(tweetInfoList);
+    public void save(List<TweetInfo> tweetInfoList)
+    {
+        for (TweetInfo tweetInfo : tweetInfoList)
+        {
+            _jongo.getCollection(_collectionName).save(tweetInfo);
+        }
     }
 
-    public void dropCollection(){
+    public MongoCursor<TweetInfo> getOrderedNumericField(String fieldName)
+    {
+        return _jongo.getCollection(_collectionName).
+                find("{}").
+                projection("{" + fieldName + ": 1, _id : 0}").
+                sort("{" + fieldName + ": 1}").
+                as(TweetInfo.class);
+    }
+
+    public void dropCollection()
+    {
         _jongo.getCollection(_collectionName).drop();
     }
 
-    public void closeMongo(){
+    public void closeMongo()
+    {
         _mongoClient.close();
     }
 }
